requirements.txt
langchain_core
langchain_openai
python-dotenv
streamlit
langchain_community
